feat(session-live): #2788 game-agnostic FlavorRenderer, wire Wingspan
FlavorRenderer no longer knows about Catan-specific label shapes: it now
takes a game-agnostic FlavorProps (session/viewerRole/sessionId/className/
livePoints/phaseName) and dispatches to a per-game lazy component via
FLAVOR_MAP, which now includes wingspan alongside catan. CatanLiveFlavor
self-builds its CatanLiveFlavorLabels internally via useTranslation/useIntl
(the exact object literal moved verbatim out of SessionLiveView), so
SessionLiveView no longer needs to construct or thread a labels prop through
either FlavorRenderer call site. Adds pages.sessionLive.flavor.wingspan.* to
it.json/en.json. Updates FlavorRenderer.test.tsx and CatanLiveFlavor.test.tsx
to render through IntlProvider now that labels are self-built.
