+++ apps/web/src/app/(authenticated)/sessions/[id]/live/_components/SessionLiveView.tsx
@@ -104,7 +104,6 @@ import {
   LiveSessionNotes,
   RightColumnTabs,
   ToolkitRenderer,
-  type CatanLiveFlavorLabels,
   type ConnectionLostBannerLabels,
   type LiveAgentChatLabels,
   type LiveSessionNotesLabels,
@@ -1131,46 +1130,6 @@ export function SessionLiveView(): ReactElement {
   const phasesQuery = useLiveSessionPhases(sessionId ?? '', showFlavorTab && sessionId != null);
   const catanPhaseName = phasesQuery.data?.currentPhaseName ?? null;
 
-  // Placeholder-bearing templates ({n}/{name}/{score}) are read RAW from
-  // intl.messages so react-intl does NOT ICU-interpolate them — the flavor
-  // component does the runtime .replace. Same pattern as the toolkitRenderer
-  // aria templates above. Non-placeholder labels use t() normally.
-  const catanFlavorLabels = useMemo<CatanLiveFlavorLabels>(
-    () => ({
-      panelAriaLabel: t('pages.sessionLive.flavor.catan.panelAriaLabel'),
-      roundTemplate:
-        (intl.messages['pages.sessionLive.flavor.catan.roundTemplate'] as string) ?? 'Round {n}',
-      activePlayerTemplate:
-        (intl.messages['pages.sessionLive.flavor.catan.activePlayerTemplate'] as string) ??
-        'Turno di {name}',
-      phaseTemplate:
-        (intl.messages['pages.sessionLive.flavor.catan.phaseTemplate'] as string) ?? 'Fase: {name}',
-      initBoardCta: t('pages.sessionLive.flavor.catan.initBoardCta'),
-      viewerWaiting: t('pages.sessionLive.flavor.catan.viewerWaiting'),
-      hexAriaTemplate:
-        (intl.messages['pages.sessionLive.flavor.catan.hexAriaTemplate'] as string) ??
-        '{terrain} {number}',
-      robberLabel: t('pages.sessionLive.flavor.catan.robberLabel'),
-      diceLastLabel: t('pages.sessionLive.flavor.catan.diceLastLabel'),
-      diceHistoryLabel: t('pages.sessionLive.flavor.catan.diceHistoryLabel'),
-      rollAriaTemplate:
-        (intl.messages['pages.sessionLive.flavor.catan.rollAriaTemplate'] as string) ??
-        'Registra tiro {n}',
-      vpLabel: t('pages.sessionLive.flavor.catan.vpLabel'),
-      handLabel: t('pages.sessionLive.flavor.catan.handLabel'),
-      devLabel: t('pages.sessionLive.flavor.catan.devLabel'),
-      settlementsLabel: t('pages.sessionLive.flavor.catan.settlementsLabel'),
-      citiesLabel: t('pages.sessionLive.flavor.catan.citiesLabel'),
-      roadsLabel: t('pages.sessionLive.flavor.catan.roadsLabel'),
-      longestRoadLabel: t('pages.sessionLive.flavor.catan.longestRoadLabel'),
-      largestArmyLabel: t('pages.sessionLive.flavor.catan.largestArmyLabel'),
-      incAriaTemplate:
-        (intl.messages['pages.sessionLive.flavor.catan.incAriaTemplate'] as string) ?? '{field} +1',
-      decAriaTemplate:
-        (intl.messages['pages.sessionLive.flavor.catan.decAriaTemplate'] as string) ?? '{field} -1',
-    }),
-    [t, intl.messages]
-  );
   // #2787: never strand the user on ?tab=flavor / ?mtab=flavor when the game has
   // no flavor (e.g. a stale bookmark carried to a non-catan session) — the flavor
   // tab button is hidden (showFlavorTab=false), so fall the panel back to 'score'.
@@ -1410,7 +1369,6 @@ export function SessionLiveView(): ReactElement {
             gameSlug={liveSessionDto.gameSlug}
             view="live"
             session={liveSessionDto}
-            labels={catanFlavorLabels}
             viewerRole={activeSession.viewerRole}
             sessionId={liveSessionDto.id}
             livePoints={catanLivePoints}
@@ -1495,7 +1453,6 @@ export function SessionLiveView(): ReactElement {
     effectiveMobileTab,
     activeSession,
     liveSessionDto,
-    catanFlavorLabels,
     catanLivePoints,
     catanPhaseName,
     sessionId,
@@ -1636,7 +1593,6 @@ export function SessionLiveView(): ReactElement {
           gameSlug={liveSessionDto.gameSlug}
           view="live"
           session={liveSessionDto}
-          labels={catanFlavorLabels}
           viewerRole={activeSession.viewerRole}
           sessionId={liveSessionDto.id}
           livePoints={catanLivePoints}
