+++ apps/web/src/components/features/session-live/__tests__/FlavorRenderer.test.tsx
@@ -1,38 +1,28 @@
 import { render, screen } from '@testing-library/react';
+import { IntlProvider } from 'react-intl';
 import { describe, expect, it, vi } from 'vitest';
 
 import type { LiveSessionDto } from '@/lib/api/schemas/live-sessions.schemas';
 
-import type { CatanLiveFlavorLabels } from '../flavors/catan/CatanLiveFlavor';
 import { FlavorRenderer, hasFlavor } from '../FlavorRenderer';
 
 vi.mock('@/hooks/mutations/useUpdateLiveGameState', () => ({
   useUpdateLiveGameState: () => ({ mutate: vi.fn() }),
 }));
 
-const LABELS: CatanLiveFlavorLabels = {
-  panelAriaLabel: 'Pannello Catan',
-  roundTemplate: 'Round {n}',
-  activePlayerTemplate: 'Turno di {name}',
-  phaseTemplate: 'Fase: {name}',
-  initBoardCta: 'Genera board Catan',
-  viewerWaiting: 'In attesa dell’host',
-  hexAriaTemplate: '{terrain} {number}',
-  robberLabel: 'Ladro',
-  diceLastLabel: 'Ultimo tiro',
-  diceHistoryLabel: 'Cronologia',
-  rollAriaTemplate: 'Registra tiro {n}',
-  vpLabel: 'PV',
-  handLabel: 'Mano',
-  devLabel: 'Sviluppo',
-  settlementsLabel: 'Insediamenti',
-  citiesLabel: 'Città',
-  roadsLabel: 'Strade',
-  longestRoadLabel: 'Strada+',
-  largestArmyLabel: 'Armata+',
-  incAriaTemplate: '{field} +1',
-  decAriaTemplate: '{field} -1',
-};
+// #2788: flavor components self-build labels via useTranslation/useIntl now,
+// so tests render through an IntlProvider. `onError` swallows react-intl
+// MISSING_TRANSLATION noise (empty messages → t() returns the raw key).
+const PANEL_ARIA_KEY = 'pages.sessionLive.flavor.catan.panelAriaLabel';
+const WAITING_KEY = 'pages.sessionLive.flavor.catan.viewerWaiting';
+
+function renderWithIntl(ui: React.ReactElement) {
+  return render(
+    <IntlProvider locale="en" messages={{}} onError={() => {}}>
+      {ui}
+    </IntlProvider>
+  );
+}
 
 const SESSION = {
   gameSlug: 'catan',
@@ -58,9 +48,10 @@ const SESSION = {
 } as unknown as LiveSessionDto;
 
 describe('hasFlavor', () => {
-  it('is true for catan, false for unknown / null / undefined', () => {
+  it('is true for catan and wingspan, false for unknown / null / undefined', () => {
     expect(hasFlavor('catan')).toBe(true);
-    expect(hasFlavor('wingspan')).toBe(false);
+    expect(hasFlavor('wingspan')).toBe(true);
+    expect(hasFlavor('chess')).toBe(false);
     expect(hasFlavor(null)).toBe(false);
     expect(hasFlavor(undefined)).toBe(false);
   });
@@ -73,7 +64,6 @@ describe('FlavorRenderer', () => {
         gameSlug="chess"
         view="live"
         session={SESSION}
-        labels={LABELS}
         viewerRole="Player"
         sessionId="s1"
       />
@@ -82,20 +72,19 @@ describe('FlavorRenderer', () => {
   });
 
   it('lazy-loads and renders the Catan flavor for gameSlug=catan, forwarding viewerRole/sessionId', async () => {
-    render(
+    renderWithIntl(
       <FlavorRenderer
         gameSlug="catan"
         view="live"
         session={SESSION}
-        labels={LABELS}
         viewerRole="Player"
         sessionId="s1"
       />
     );
-    // <section aria-label="Pannello Catan"> → implicit role="region"
-    expect(await screen.findByRole('region', { name: 'Pannello Catan' })).toBeInTheDocument();
+    // <section aria-label={t(...)}> → implicit role="region"
+    expect(await screen.findByRole('region', { name: PANEL_ARIA_KEY })).toBeInTheDocument();
     // Player (non-Host) sees the waiting message, not the host CTA — proves
     // viewerRole was actually forwarded through to the lazy flavor.
-    expect(screen.getByText('In attesa dell’host')).toBeInTheDocument();
+    expect(screen.getByText(WAITING_KEY)).toBeInTheDocument();
   });
 });
