+++ apps/web/src/components/features/session-live/index.ts
@@ -143,16 +143,14 @@ export type {
   RightColumnTabsProps,
 } from '@/components/features/session-live/RightColumnTabs';
 
-// ─── G6a #2787 Catan flavor (ADR-070 lazy per-game modules) ───────────────────
+// ─── G6a #2787 / #2788 per-game flavor registry (ADR-070 lazy per-game modules) ─
 export { FlavorRenderer, hasFlavor } from '@/components/features/session-live/FlavorRenderer';
 export type {
+  FlavorProps,
   FlavorRendererProps,
   FlavorView,
 } from '@/components/features/session-live/FlavorRenderer';
-export type {
-  CatanLiveFlavorLabels,
-  CatanLiveFlavorProps,
-} from '@/components/features/session-live/flavors/catan/CatanLiveFlavor';
+export type { CatanLiveFlavorProps } from '@/components/features/session-live/flavors/catan/CatanLiveFlavor';
 
 export { SessionToolsRail } from '@/components/features/session-live/SessionToolsRail';
 export type {
